requirements.txt
python-telegram-bot==20.7
ollama==0.1.9
flask==2.3.3
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
langgraph==0.2.0
langchain==0.3.0
langchain-core==0.3.0
typing-extensions==4.12.0
pydantic==2.0.0
Pillow==10.1.0
opencv-python==4.9.0
numpy==1.24.0
transformers==4.38.0
torch==2.2.0
torchvision==0.17.0
faiss-cpu==1.7.4
h5py==3.10.0
tqdm==4.66.0
psutil==5.9.0
python-dateutil==2.8.2
hydra-core==1.3.2
omegaconf==2.3.0
jupyter==1.0.0
ipython==8.18.0